Refactor request creation
