+++ tekore/client/base.py
@@ -1,13 +1,25 @@
 import json
 
-from requests import Request, HTTPError
+from requests import Request, Response, HTTPError
 
 from tekore.sender import Sender, Client
 from tekore.model.error import PlayerErrorReason
 
+prefix = 'https://api.spotify.com/v1/'
 error_format = """Error in {url}:
 {code}: {msg}
 """
+
+
+def build_url(url: str) -> str:
+    if not url.startswith('http'):
+        url = prefix + url
+    return url
+
+
+def parse_url_params(params: dict = None) -> dict:
+    params = params or {}
+    return {k: v for k, v in params.items() if v is not None}
 
 
 def parse_json(response):
@@ -29,9 +41,17 @@ def parse_error_reason(response):
     return reason
 
 
-class SpotifyBase(Client):
-    prefix = 'https://api.spotify.com/v1/'
+def handle_errors(request: Request, response: Response) -> None:
+    if response.status_code >= 400:
+        error_str = error_format.format(
+            url=response.url,
+            code=response.status_code,
+            msg=parse_error_reason(response)
+        )
+        raise HTTPError(error_str, request=request, response=response)
 
+
+class SpotifyBase(Client):
     def __init__(
             self,
             token=None,
@@ -50,37 +70,11 @@ class SpotifyBase(Client):
         super().__init__(sender)
         self.token = token
 
-    def _build_request(self, method: str, url: str, headers: dict = None) -> Request:
-        if not url.startswith('http'):
-            url = self.prefix + url
-
-        default_headers = {
+    def _create_headers(self, content_type: str = 'application/json'):
+        return {
             'Authorization': f'Bearer {str(self.token)}',
-            'Content-Type': 'application/json'
+            'Content-Type': content_type
         }
-        default_headers.update(headers or {})
-
-        return Request(method, url, headers=default_headers)
-
-    @staticmethod
-    def _set_content(request: Request, payload=None, params: dict = None) -> None:
-        params = params or {}
-        request.params = {k: v for k, v in params.items() if v is not None}
-        if payload is not None:
-            if request.headers['Content-Type'] == 'application/json':
-                request.data = json.dumps(payload)
-            else:
-                request.data = payload
-
-    @staticmethod
-    def _handle_errors(request, response) -> None:
-        if response.status_code >= 400:
-            error_str = error_format.format(
-                url=response.url,
-                code=response.status_code,
-                msg=parse_error_reason(response)
-            )
-            raise HTTPError(error_str, request=request, response=response)
 
     def _request(
             self,
@@ -89,10 +83,15 @@ class SpotifyBase(Client):
             payload=None,
             params: dict = None
     ):
-        request = self._build_request(method, url)
-        self._set_content(request, payload, params)
+        request = Request(
+            method=method,
+            url=build_url(url),
+            headers=self._create_headers(),
+            params=parse_url_params(params),
+            data=json.dumps(payload) if payload is not None else None
+        )
         response = self._send(request)
-        self._handle_errors(request, response)
+        handle_errors(request, response)
         return parse_json(response)
 
     def _get(self, url: str, payload=None, **params):
@@ -106,13 +105,3 @@ class SpotifyBase(Client):
 
     def _put(self, url: str, payload=None, **params):
         return self._request('PUT', url, payload=payload, params=params)
-
-    def _get_paging_result(self, address: str):
-        result = self._get(address)
-
-        # If only one top-level key, the paging object is one level deeper
-        if len(result) == 1:
-            key = list(result.keys())[0]
-            result = result[key]
-
-        return result
